+++ src/main/java/Event.java
@@ -1,0 +1,25 @@
+public class Event extends Task {
+    protected String from, to;
+
+    public Event(String description, String from, String to) {
+        super(description);
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+    }
+
+    @Override
+    public String typeTag() {
+        return "E";
+    }
+
+    @Override
+    public String toDataString() {
+        return String.format("%s | %d | %s | %s | %s",
+                typeTag(), isDone ? 1 : 0, esc(description), esc(from), esc(to));
+    }
+}
