+++ src/main/java/PipException.java
@@ -1,0 +1,5 @@
+public class PipException extends Exception {
+    public PipException(String message) {
+        super(message);
+    }
+}
